services/gtfs_status/requirements.txt
django==4.0.2
google-cloud-storage==2.1.0
